before:
```
Finished in 12.19 seconds (files took 7.11 seconds to load)
18 examples, 0 failures
```
after:
```
Finished in 3.49 seconds (files took 7.23 seconds to load)
18 examples, 0 failures
```

There are still some DB writes in the test, but most are avoided.